+++ lib/Container.dart
@@ -7,21 +7,3 @@ abstract class Container {
 
   bool AddLiquid(Liquid liquid);
 }
-
-class Cup implements Container{
-  @override
-  String name;
-
-  @override
-  double size;
-
-  @override
-  List<Container> usualContainers;
-
-  @override
-  bool AddLiquid(Liquid liquid) {
-    // TODO: implement AddLiquid
-    throw UnimplementedError();
-  }
-
-}
